+++ EssentialFeed/EssentialFeed/FeedCache/FeedStore.swift
@@ -1,0 +1,17 @@
+//
+//  FeedStore.swift
+//  EssentialFeed
+//
+//  Created by Aaron Huánuco on 13/09/2020.
+//  Copyright © 2020 Aaron Huánuco. All rights reserved.
+//
+
+import Foundation
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
